fix tests and migration

+++ src/main/resources/db/changelog/db.changelog-1.0.sql
@@ -3,7 +3,7 @@
 -- changeset kostapchuk:1
 CREATE TABLE IF NOT EXISTS users
 (
-    id            SERIAL PRIMARY KEY,
+    id            BIGINT PRIMARY KEY,
     serial        VARCHAR(255) NOT NULL UNIQUE,
     document_type VARCHAR(50)  NOT NULL
 );
@@ -12,7 +12,7 @@ CREATE TABLE IF NOT EXISTS users
 -- changeset kostapchuk:2
 CREATE TABLE IF NOT EXISTS account
 (
-    id       SERIAL PRIMARY KEY,
+    id       BIGINT PRIMARY KEY,
     balance  DECIMAL(19, 2) NOT NULL,
     user_id  BIGINT REFERENCES users (id),
     currency VARCHAR(50)    NOT NULL
@@ -22,7 +22,7 @@ CREATE TABLE IF NOT EXISTS account
 -- changeset kostapchuk:3
 CREATE TABLE IF NOT EXISTS operation
 (
-    id          SERIAL PRIMARY KEY,
+    id          BIGINT PRIMARY KEY,
     sender_id   BIGINT REFERENCES account (id),
     receiver_id BIGINT REFERENCES account (id),
     amount      DECIMAL(19, 2) NOT NULL,
